Department of Homeland Security Match Counter
The US Department of Homeland Security needs to determine potential
unwanted individuals at the next United Nations conference scheduled in
New York City. The US office received an encrypted list of
ill-intentioned people who may want to show up to the conference and
needs to be able to match the names on this list to the names of actual
attendees. The problem that they are facing is that they managed to only
partially decrypt the names of potential troublemakers. The decryption
process is taking a long time and the conference is about to begin. This
program matches the list of partially decrypted names to the names of
actual attendees and generates a report that shows matching percentages
between the names on both lists.